% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_profile_range.R
\name{check_profile_range}
\alias{check_profile_range}
\title{Check the profile range specified}
\usage{
check_profile_range(mydir, model_settings)
}
\arguments{
\item{mydir}{Directory where model files are located.
There is no default entry for \code{mydir}.
An example of user input could be
\code{mydir = file.path("C:/my_models", "base_model")}.}

\item{model_settings}{A list of all model settings.
Typically, created using \link{get_settings} but users can create their own list,
as long as it has all of the necessary components for the function at hand.}
}
\value{
Nothing is explicitly returned from \code{check_profile_range()}
}
\description{
Check the profile range specified
}
\author{
Chantel Wetzel
}
