% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jitter_quants.R
\name{get_jitter_quants}
\alias{get_jitter_quants}
\title{Create output files with jitter quantities}
\usage{
get_jitter_quants(mydir, model_settings, output)
}
\arguments{
\item{mydir}{Directory where model files are located.
There is no default entry for \code{mydir}.
An example of user input could be
\code{mydir = file.path("C:/my_models", "base_model")}.}

\item{model_settings}{A list of all model settings.
Typically, created using \link{get_settings} but users can create their own list,
as long as it has all of the necessary components for the function at hand.}

\item{output}{List of model output created by \code{\link[=run_jitter]{run_jitter()}}.}
}
\value{
Nothing is explicitly returned from \code{\link[=get_jitter_quants]{get_jitter_quants()}}.
}
\description{
Tables generated to visualize results
}
\author{
Chantel Wetzel
}
