% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_retro_quants.R
\name{get_retro_quants}
\alias{get_retro_quants}
\title{Get model values across retrospective runs}
\usage{
get_retro_quants(mydir, model_settings, output)
}
\arguments{
\item{mydir}{Directory where model files are located.
There is no default entry for \code{mydir}.
An example of user input could be
\code{mydir = file.path("C:/my_models", "base_model")}.}

\item{model_settings}{A list of all model settings.
Typically, created using \link{get_settings} but users can create their own list,
as long as it has all of the necessary components for the function at hand.}

\item{output}{List of model output created by \code{\link[=run_retro]{run_retro()}}.}
}
\value{
Nothing is explicitly returned from \code{\link[=get_retro_quants]{get_retro_quants()}}.
The following objects are saved to the disk.
\itemize{
\item \code{mohnsrho.csv} with the following columns:
\itemize{
\item type: the type of Mohn's rho
\itemize{
\item \href{https://academic.oup.com/icesjms/article/56/4/473/624639}{Mohn (1999)},
\item Woods Hole Mohn's rho \href{https://archive.nefmc.org/tech/council_mtg_docs/Sept\%202009/Herring/Doc\%209_Retro\%20Working\%20Group\%20Report.pdf}{(Legault 2009)} used by the \href{https://www.fisheries.noaa.gov/about/northeast-fisheries-science-center}{Northeast Fisheries Science Center (NEFSC)}, and
\item \href{https://doi.org/10.1093/icesjms/fsu198}{Hurtado-Ferro et al. (2015)} used by the \href{https://www.fisheries.noaa.gov/about/alaska-fisheries-science-center}{Alaska Fisheries Science Center (AFSC)}
}
\item Quantity: the stock assessment quantity of interest
\item values: the Mohn's rho values
}

\code{apply(utils::read.csv(file.path("..", paste0(mod_loc, "_retro"), "retrofigures4doc.csv")), 1, function(x) do.call(sa4ss::add_figure, as.list(x)))}
\item \code{mohnsrho.tex} for use with \code{sa4ss::read_child()}
inside of an environment with \code{results = "asis"}
to include a table of Mohn's rho values in a document.

\code{sa4ss::read_child(file.path(paste0(params[["model"]], "_retro"), "mohnsrho.tex"))}
}
}
\description{
Get model values across retrospective runs
}
\author{
Chantel Wetzel
}
