% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_settings.R
\name{get_settings}
\alias{get_settings}
\title{Check that all of the settings are in the list}
\usage{
get_settings(mydir = NULL, settings = NULL)
}
\arguments{
\item{mydir}{Directory where model files are located.
There is no default entry for \code{mydir}.
An example of user input could be
\code{mydir = file.path("C:/my_models", "base_model")}.}

\item{settings}{A list of the current settings where each object in the list
must be named. Those names that are not found in the stored list will be
added. The default value of \code{NULL} leads to a full list being
returned.}
}
\value{
A list of settings for running model diagnostics.
}
\description{
Check that all of the settings are in the list
}
\examples{
\dontrun{
get_settings(
  mydir = "directory",
  settings = list("Njitter" = 10)
)
}

}
\author{
Chantel Wetzel & Kelli Faye Johnson
}
