% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_summary.R
\name{get_summary}
\alias{get_summary}
\title{Generate likelihood profiles
To be called from the run_diagnostics function after creating
the model settings using the get_settings function.}
\usage{
get_summary(mydir, para, vec, name, profilemodels, profilesummary)
}
\arguments{
\item{mydir}{Directory where model files are located.
There is no default entry for \code{mydir}.
An example of user input could be
\code{mydir = file.path("C:/my_models", "base_model")}.}

\item{para}{SS parameter name that the profile was run across}

\item{vec}{vector of parameter values the profile covers}

\item{name}{Identify if the csv file show jitter, retro, or profile results}

\item{profilemodels}{object created by the SSgetoutput funciton}

\item{profilesummary}{object created by the SSsummarize funciton}
}
\description{
Generate likelihood profiles
To be called from the run_diagnostics function after creating
the model settings using the get_settings function.
}
\author{
Chantel Wetzel & Kelli Johnson
}
