% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jitter_wrapper.R
\name{jitter_wrapper}
\alias{jitter_wrapper}
\title{Code to run jitters for a model
Output will be saved in an Rdata object called "jitter_output"
Plots and tables generated to visualize results}
\usage{
jitter_wrapper(mydir, model_settings)
}
\arguments{
\item{mydir}{Directory where model files are located.}

\item{model_settings}{input of all settings created using the get_settings function}
}
\value{
A vector of likelihoods for each jitter iteration.
}
\description{
Code to run jitters for a model
Output will be saved in an Rdata object called "jitter_output"
Plots and tables generated to visualize results
}
\author{
Chantel Wetzel
}
