% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_plot.R
\name{plot_profile}
\alias{plot_profile}
\title{Generate likelihood profile plots}
\usage{
plot_profile(mydir, rep, para, profilesummary)
}
\arguments{
\item{mydir}{Directory where model files are located.
There is no default entry for \code{mydir}.
An example of user input could be
\code{mydir = file.path("C:/my_models", "base_model")}.}

\item{rep}{A list of model output as returned by \code{\link[r4ss:SS_output]{r4ss::SS_output()}}.}

\item{para}{A character string specifying the SS3 parameter name that the
profile pertains to. The parameter name should match the name in the
control.ss_new file from SS3.
The name will be cleaned up for plotting purposes internally within the function.
For example, \code{SR_BH_steep} becomes "Steepness (\emph{h})".}

\item{profilesummary}{Output from \code{\link[r4ss:SSsummarize]{r4ss::SSsummarize()}}.
Ensure that the results are ordered according to the parameter of interest.
Otherwise, the profile plots will not be smooth and lines will zig-zag back and forth.}
}
\description{
Create plots to be included in presentations and documents that
summarize a likelihood profile.
}
\seealso{
\link{profile_wrapper} and \link{rerun_profile_vals} call \code{plot_profile}.
}
\author{
Chantel Wetzel.
}
