% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_retro.R
\name{plot_retro}
\alias{plot_retro}
\title{Create plots for retrospective runs}
\usage{
plot_retro(mydir, model_settings, output)
}
\arguments{
\item{mydir}{Directory where model files are located.
There is no default entry for \code{mydir}.
An example of user input could be
\code{mydir = file.path("C:/my_models", "base_model")}.}

\item{model_settings}{A list of all model settings.
Typically, created using \link{get_settings} but users can create their own list,
as long as it has all of the necessary components for the function at hand.}

\item{output}{List of model output created by \code{\link[=run_retro]{run_retro()}}.}
}
\value{
Nothing is explicitly returned from \code{\link[=plot_retro]{plot_retro()}}.
The following objects are saved to the disk.
\itemize{
\item A set of figures from \link[r4ss:SSplotComparisons]{r4ss::SSplotComparisons}
\item \code{retrofigures4doc.csv} for use with \code{sa4ss::add_figure} to add retro peels of
spawning stock biomass (\emph{SSB}) and fraction unfished,
complete with captions and alternative text.
}
}
\description{
Create plots for retrospective runs
}
\author{
Chantel Wetzel
}
