% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_fxn.R
\name{pngfun}
\alias{pngfun}
\title{PNG figure funtion}
\usage{
pngfun(wd, file, w = 7, h = 7, pt = 12)
}
\arguments{
\item{wd}{location to save figure}

\item{file}{plot name}

\item{w}{figure width}

\item{h}{figure height in inches}

\item{pt}{point size}
}
\description{
PNG figure funtion
}
