% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_decimal.R
\name{pretty_decimal}
\alias{pretty_decimal}
\title{Format a vector of real numbers with a pretty number of decimal places}
\usage{
pretty_decimal(x)
}
\arguments{
\item{x}{A vector of real numbers.}
}
\value{
A vector of strings containing formatted numerical values all with
an equal number of decimal places and at least one decimal place.
}
\description{
Choose the final formatting for the number of decimals that should be
displayed given a vector of real numbers. All numbers in the returned
vector will be unique.
}
\examples{
pretty_decimal(c(1.0, pi, 3.14, 0.002, 0.0021, 0.02))

# Returns at least one decimal place even if not needed
pretty_decimal(1)

# Will return an error because duplicated entries
\dontrun{
pretty_decimal(c(1.1, 1.1))
}

}
\author{
Kelli F. Johnson
}
