% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_wrapper.R
\name{profile_wrapper}
\alias{profile_wrapper}
\title{Generate likelihood profiles
To be called from the run_diagnostics function after creating
the model settings using the get_settings function.}
\usage{
profile_wrapper(mydir, model_settings)
}
\arguments{
\item{mydir}{Directory where model files are located.}

\item{model_settings}{input of all settings created using the get_settings function}
}
\description{
Generate likelihood profiles
To be called from the run_diagnostics function after creating
the model settings using the get_settings function.
}
\author{
Chantel Wetzel.
}
