% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{round_any}
\alias{round_any}
\title{Round values}
\usage{
round_any(x, accuracy, f = round)
}
\arguments{
\item{x}{Numeric value to round.}

\item{accuracy}{The number of digits to round to.}

\item{f}{R function to apply. Default is round().}
}
\description{
Round values
}
