% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_diagnostics.R
\name{run_diagnostics}
\alias{run_diagnostics}
\title{Wrapper to run each of the 3 standard diagnostic items:
1) Jitter
2) Profiles across female m, steepness, and R0
3) Retrospectives}
\usage{
run_diagnostics(mydir, model_settings)
}
\arguments{
\item{mydir}{Directory where model files are located.}

\item{model_settings}{input of all settings created using the get_settings function}
}
\value{
A vector of likelihoods for each jitter iteration.
}
\description{
Wrapper to run each of the 3 standard diagnostic items:
1) Jitter
2) Profiles across female m, steepness, and R0
3) Retrospectives
}
\author{
Chantel Wetzel
}
