% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param_values.R
\name{get_param_values}
\alias{get_param_values}
\title{Generate likelihood profiles
To be called from the run_diagnostics function after creating
the model settings using the get_settings function.}
\usage{
get_param_values(mydir, para = NULL, name = NULL, vec, summary)
}
\arguments{
\item{mydir}{Directory where model files are located.
There is no default entry for \code{mydir}.
An example of user input could be
\code{mydir = file.path("C:/my_models", "base_model")}.}

\item{para}{SS3 parameter name that the profile was run across}

\item{name}{Identify if the csv file show jitter, retro, or profile results}

\item{vec}{vector of parameter values or retrospective runs the summary object contains}

\item{summary}{object created by the SSsummarize function}
}
\description{
Generate likelihood profiles
To be called from the run_diagnostics function after creating
the model settings using the get_settings function.
}
\author{
Chantel Wetzel & Kelli Johnson
}
