% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param_values.R
\name{get_param_values}
\alias{get_param_values}
\title{Generate likelihood profiles
To be called from the run_diagnostics function after creating
the model settings using the get_settings function.}
\usage{
get_param_values(mydir, para = NULL, vec, summary)
}
\arguments{
\item{mydir}{Directory where model files are located.
There is no default entry for \code{mydir}.
An example of user input could be
\code{mydir = file.path("C:/my_models", "base_model")}.}

\item{para}{SS3 parameter name (string) that the profile was run across.}

\item{vec}{Vector of parameter values or retrospective runs the summary object contains}

\item{summary}{List created by the \code{\link[r4ss:SSsummarize]{r4ss::SSsummarize()}} function}
}
\description{
Generate likelihood profiles
To be called from the run_diagnostics function after creating
the model settings using the get_settings function.
}
\author{
Chantel Wetzel & Kelli Johnson
}
