% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_summary.R
\name{get_summary}
\alias{get_summary}
\title{Generate results from likelihood profiles}
\usage{
get_summary(mydir, para, vec, profilemodels, profilesummary)
}
\arguments{
\item{mydir}{Directory where model files are located.
There is no default entry for \code{mydir}.
An example of user input could be
\code{mydir = file.path("C:/my_models", "base_model")}.}

\item{para}{A character string specifying the SS3 parameter name that the
profile pertains to. The parameter name should match the name in the
control.ss_new file from SS3.}

\item{vec}{A numeric vector specifying the parameter values that the
profile covers.}

\item{profilemodels}{An object returned from \code{\link[r4ss:SSgetoutput]{r4ss::SSgetoutput()}}.}

\item{profilesummary}{An object returned from \code{\link[r4ss:SSsummarize]{r4ss::SSsummarize()}}.}
}
\description{
Used by \code{\link[=profile_wrapper]{profile_wrapper()}} to write summary statistics to the disk.
}
\author{
Chantel Wetzel & Kelli Johnson
}
