% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jitter_wrapper.R
\name{jitter_wrapper}
\alias{jitter_wrapper}
\title{Run \link[r4ss:jitter]{r4ss::jitter} based on \code{model_settings}}
\usage{
jitter_wrapper(mydir, model_settings)
}
\arguments{
\item{mydir}{Directory where model files are located.
There is no default entry for \code{mydir}.
An example of user input could be
\code{mydir = file.path("C:/my_models", "base_model")}.}

\item{model_settings}{A list of all model settings.
Typically, created using \link{get_settings} but users can create their own list,
as long as it has all of the necessary components for the function at hand.}
}
\value{
Nothing is explicitly returned from \code{jitter_wrapper}.
}
\description{
Code to run jitters for a model
Output will be saved in an Rdata object called "jitter_output"
Plots and tables generated to visualize results
}
\seealso{
The following functions interact with \code{jitter_wrapper}:
\itemize{
\item \link{run_diagnostics}: calls \code{jitter_wrapper}
}
}
\author{
Chantel Wetzel
}
