% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jitter_wrapper.R
\name{jitter_wrapper}
\alias{jitter_wrapper}
\title{Code to run jitters for a model
Output will be saved in an Rdata object called "jitter_output"
Plots and tables generated to visualize results}
\usage{
jitter_wrapper(mydir, model_settings)
}
\arguments{
\item{mydir}{Directory where model files are located.
There is no default entry for `mydir`.
An example of user input could be
`mydir = file.path("C:/my_models", "base_model")`.}

\item{model_settings}{A list of all model settings.
Typically, created using [get_settings] but users can create their own list,
as long as it has all of the necessary components for the function at hand.}
}
\value{
A vector of likelihoods for each jitter iteration.
}
\description{
Code to run jitters for a model
Output will be saved in an Rdata object called "jitter_output"
Plots and tables generated to visualize results
}
\author{
Chantel Wetzel
}
