% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_fxn.R
\name{pngfun}
\alias{pngfun}
\title{Create a \code{.png} figure}
\usage{
pngfun(wd, file, w = 7, h = 7, pt = 12)
}
\arguments{
\item{wd}{A string providing the directory location where you want the
figure saved.}

\item{file}{A string providing the name of the .png file that will be
saved in \code{wd}.}

\item{w, h}{Numeric values providing the figure width and height in inches.}

\item{pt}{Numeric value providing the point size.}
}
\description{
Create a \code{.png} figure
}
\author{
by Chantel Wetzel
}
