% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_plot.R
\name{profile_plot}
\alias{profile_plot}
\title{Generate likelihood profiles plots
To be called from the run_diagnostics function after creating
the model settings using the get_settings function.}
\usage{
profile_plot(mydir, model_settings, rep, vec, para, profilesummary)
}
\arguments{
\item{mydir}{Directory where model files are located.
There is no default entry for `mydir`.
An example of user input could be
`mydir = file.path("C:/my_models", "base_model")`.}

\item{model_settings}{A list of all model settings.
Typically, created using [get_settings] but users can create their own list,
as long as it has all of the necessary components for the function at hand.}

\item{rep}{base model outpus}

\item{vec}{vector of value that were profiled over}

\item{para}{parameter name SS control.ss_new expected parameter name}

\item{profilesummary}{output from [r4ss::SSsummarize()]}
}
\description{
Generate likelihood profiles plots
To be called from the run_diagnostics function after creating
the model settings using the get_settings function.
}
\author{
Chantel Wetzel.
}
