% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_wrapper.R
\name{profile_wrapper}
\alias{profile_wrapper}
\title{Generate likelihood profiles
To be called from the run_diagnostics function after creating
the model settings using the get_settings function.}
\usage{
profile_wrapper(mydir, model_settings)
}
\arguments{
\item{mydir}{Directory where model files are located.
There is no default entry for `mydir`.
An example of user input could be
`mydir = file.path("C:/my_models", "base_model")`.}

\item{model_settings}{A list of all model settings.
Typically, created using [get_settings] but users can create their own list,
as long as it has all of the necessary components for the function at hand.}
}
\description{
Generate likelihood profiles
To be called from the run_diagnostics function after creating
the model settings using the get_settings function.
}
\author{
Chantel Wetzel.
}
