% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retro_wrapper.R
\name{retro_wrapper}
\alias{retro_wrapper}
\title{Retrospecitive runs
To be called from the run_diagnostics function after creating
the model settings using the get_settings function.}
\usage{
retro_wrapper(mydir, model_settings)
}
\arguments{
\item{mydir}{Directory where model files are located.}

\item{model_settings}{input of all settings created using the get_settings function}
}
\description{
Retrospecitive runs
To be called from the run_diagnostics function after creating
the model settings using the get_settings function.
}
\author{
Chantel Wetzel.
}
